% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns.R
\name{dns_helper}
\alias{dns_helper}
\title{helper function for Dynamic Nelson-Siegel Model}
\usage{
dns_helper(maturity, lambda)
}
\arguments{
\item{maturity}{vector with maturities in months}

\item{lambda}{the lambda}
}
\description{
helper function for Dynamic Nelson-Siegel Model
}
