% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns.R
\name{estimDNS}
\alias{estimDNS}
\title{Estimate Dynamic Nelson-Siegel Model}
\usage{
estimDNS(yields, maturity, lambda, frequency = 12)
}
\arguments{
\item{yields}{ts-object with yields}

\item{maturity}{vector with maturities in years}

\item{lambda}{parameter in Nelson-Siegel}

\item{frequency}{frequency of the data (12 = monthly, 52 = weekly, 360 = daily)}
}
\description{
Estimate Dynamic Nelson-Siegel Model
}
