% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dss.r
\name{estimDSS}
\alias{estimDSS}
\title{Dynamic Svensson-Söderlind Model}
\usage{
estimDSS(yields, maturity, lambda1, lambda2, frequency = 12)
}
\arguments{
\item{yields}{Txn series of yields}

\item{maturity}{vector of length n of maturity dates}

\item{lambda1, lambda2}{decay parameters}

\item{frequency}{frequency of the data}
}
\description{
Dynamic Svensson-Söderlind Model
}
