% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estim.R
\name{EstimYieldCurve}
\alias{EstimYieldCurve}
\title{Estimate Yield Curve Models}
\usage{
EstimYieldCurve(yields, exogen = NULL, maturity, frequency = 12,
  method = "DNS")
}
\arguments{
\item{yields}{A Txn panel of yields. Has to be a ts-objects}

\item{exogen}{Exogeneous Variables for the VAR-Model}

\item{maturity}{An nx1-vector of maturities.}

\item{frequency}{data frequency (12=monthly, 52 = weekly, 360 = daily)}

\item{method}{The method which is used to extract the factors. (DNS = Dynamic Nelson Siegel)}
}
\description{
This function estimates yield curve models and computes term premia. Users have the choice between several yield curve models such as the Dynamic-Nelson-Siegel model introduced by (), the Dynamic Svensson-Söderlind model or Joslin-Singleton-Zhu-model. For the DNS and DSS model the parameter lambda is automatically computed using a grid search.
}
