% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/premium.r
\name{Expr}
\alias{Expr}
\title{Function to calculate the risk-free rate (Expectations component)}
\usage{
Expr(V, D, n, X, mp, rho, I)
}
\arguments{
\item{V}{eigenvectors of VAR-coefficient}

\item{D}{eigenvalues of VAR-coefficient}

\item{n}{maturity in months}

\item{X}{current yield factors}

\item{mp}{mean value of factors}

\item{rho}{helper}

\item{I}{identity matrix}
}
\description{
Function to calculate the risk-free rate (Expectations component)
}
