% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns.R
\name{find_lambda}
\alias{find_lambda}
\title{Find lambda that minimizes the sum of squared errors using a grid search}
\usage{
find_lambda(yields, maturity)
}
\arguments{
\item{yields}{txn-panel with yields}

\item{maturity}{vector with maturities in months}
}
\description{
Find lambda that minimizes the sum of squared errors using a grid search
}
