% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast.yldcurve}
\alias{forecast.yldcurve}
\title{Forecasting yield curves}
\usage{
\method{forecast}{yldcurve}(estimObj, nhor = 12, ylddata = NULL,
  maturity = NULL)
}
\arguments{
\item{estimObj}{yield curve model estimated by EstimYieldCurve}

\item{nhor}{forecast horizon}

\item{ylddata}{additional data to compute forecast errors. If NULL, no forecast errors are computed. If not NULL the user has to provide the maturity of the yields. Users should provide ylddata for the full forecast horizon.}

\item{maturity}{maturity for forecasted yields. If not provided maturity data from model will be used. Not optional if ylddata is not NULL.}
}
\description{
Forecasting yield curves
}
