% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_errors}
\alias{plot_errors}
\alias{plot_errors.yldcurve}
\title{plot fit errors of fitted model}
\usage{
plot_errors(estimObj, ...)

\method{plot_errors}{yldcurve}(estimObj, ...)
}
\arguments{
\item{estimObj}{an estimated yield curve model}

\item{...}{not used}
}
\description{
plot fit errors of fitted model
}
