% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_factors}
\alias{plot_factors}
\alias{plot_factors.yldcurve}
\title{plots estimated factor series}
\usage{
plot_factors(estimObj, ...)

\method{plot_factors}{yldcurve}(estimObj, ...)
}
\arguments{
\item{estimObj}{an estimated yield curve model}

\item{...}{not used}
}
\description{
plots estimated factor series
}
