% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_premia}
\alias{plot_premia}
\alias{plot_premia.yldcurve}
\title{plot term premia}
\usage{
plot_premia(estimObj, ...)

\method{plot_premia}{yldcurve}(estimObj, ...)
}
\arguments{
\item{estimObj}{estimated and fitted Model}

\item{...}{currently not used}
}
\description{
plot term premia
}
