% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/premium.r
\name{TermPremium}
\alias{TermPremium}
\title{Calculate Term Premia}
\usage{
TermPremium(objVar, objModel)
}
\arguments{
\item{objVar}{estimate varm-model}

\item{objModel}{estimated Model}
}
\description{
Calculate Term Premia
}
