% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{US_Yield_Curve}
\alias{US_Yield_Curve}
\title{US-Yield curves}
\format{An object of class \code{xts} (inherits from \code{zoo}) with 372 rows and 30 columns.}
\usage{
US_Yield_Curve
}
\description{
US-Yield curves
}
\keyword{datasets}
