% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varm.r
\name{varm}
\alias{varm}
\title{Estimate a VAR(1)-Model
The function varm estimates the following VAR(1)-Model:
Y_t = mp + beta * (Y_{t-1} - mp) +e_t}
\usage{
varm(y, exo = NULL)
}
\arguments{
\item{y}{data}

\item{exo}{exogeneous variables}
}
\description{
Estimate a VAR(1)-Model
The function varm estimates the following VAR(1)-Model:
Y_t = mp + beta * (Y_{t-1} - mp) +e_t
}
