% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_errors}
\alias{plot_errors}
\alias{plot_errors.yldcurve}
\alias{plot_errors.fcylds}
\title{plot errors of fitted model or forecast errors}
\usage{
plot_errors(estimObj, ...)

\method{plot_errors}{yldcurve}(estimObj, ...)

\method{plot_errors}{fcylds}(estimObj, ...)
}
\arguments{
\item{estimObj}{an estimated yield curve model}

\item{...}{not used}
}
\description{
plot errors of fitted model or forecast errors
}
